requirements.txt
requests>=2.31.0
pandas>=2.0.0
google-play-scraper>=1.2.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# NLP & Text Processing
emoji>=2.0.0
konlpy>=0.6.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
datasets>=2.12.0
nltk>=3.8.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0

pytest>=7.0.0

# python-dotenv>=1.0.0        # .env 파일 로드
# setuptools>=68.0.0           # 패키지 설치용
# pytest>=7.0.0               # 테스트용
# jupyter>=1.0.0               # 분석용 노트북
# matplotlib>=3.7.0            # 시각화용
# seaborn>=0.12.0              # 고급 시각화용
